requirements.txt
asn1crypto==0.24.0
boto3==1.6.3
botocore==1.9.3
cffi==1.10.0
cryptography==1.8.1
Django==2.0
django-cors-headers==2.1.0
django-crispy-forms==1.6.1
django-storages==1.6.5
django-tinymce==2.6.0
djangorestframework==3.7.3
docutils==0.14
idna==2.6
jmespath==0.9.3
olefile==0.44
Pillow==4.3.0
psycopg2==2.7.4
pycparser==2.18
PyJWT==1.5.3
pyOpenSSL==17.5.0
python-dateutil==2.6.1
pytz==2017.3
s3transfer==0.1.12
six==1.11.0
